requirements.txt
# Core Django and environment
django==4.2.7
python-dotenv==1.0.0
requests==2.31.0

# LLM APIs
openai==1.3.0
anthropic==0.3.11
groq==0.4.1

# Google Services
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2

# RAG Components
langchain==0.1.0
chromadb==0.4.22
duckduckgo-search==4.1.1
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.26.2